backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.115.0
google-api-core==2.30.1
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.73.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.27.2
idna==3.11
langdetect
language_tool_python==2.8.1
lxml==5.3.0
packaging==26.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.3.2
psutil
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.38.6
tqdm==4.67.3
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
